Added the actual S&W files. Shoulda done this earlier
