--- JoJoStands.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.1.32421.90
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JoJoStands", "JoJoStands.csproj", "{6D93A6BA-6527-457C-A820-7F2C839CC46F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6D93A6BA-6527-457C-A820-7F2C839CC46F}.Debug|Any CPU.ActiveCfg = Debug|ARM64
-		{6D93A6BA-6527-457C-A820-7F2C839CC46F}.Debug|Any CPU.Build.0 = Debug|ARM64
-		{6D93A6BA-6527-457C-A820-7F2C839CC46F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6D93A6BA-6527-457C-A820-7F2C839CC46F}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {C7647D74-6B87-46D3-ADB2-595BCE4A7DC9}
-	EndGlobalSection
-EndGlobal
