+++ Projectiles/PlayerStands/SoftAndWet/SoftAndWet_Idle.png
